Implement the PHP password API for hashes
This makes the PHP password API the default/latest hash version to use.
The password API will be used unless salt is set to false (the default) and fallback to the previous default of SHA256.
The hash validation will _only_ check for the expected hash version.
Implements a new EncryptorInterface method to allow for testing if the hash needs to be rehashed.
The default implementation checks for the latest hash version and uses the password API for checking password hashes.
Tests have been updated with the password API expectations and to test the needsRehash method.